Issue 2311: runtime error: index out of range [0] with length 0 while extracting state of the csrf
