+++ oauthproxy.go
@@ -788,12 +788,8 @@ func (p *OAuthProxy) doOAuthStart(rw http.ResponseWriter, req *http.Request, ove
 
 	callbackRedirect := p.getOAuthRedirectURI(req)
 	hashOAuthState := csrf.HashOAuthState()
-	loginURL := p.provider.GetLoginURL(
-		callbackRedirect,
-		encodeState(hashOAuthState, appRedirect),
-		csrf.HashOIDCNonce(),
-		extraParams,
-	)
+	loginURL := p.provider.GetLoginURL(callbackRedirect, encodeState(hashOAuthState, appRedirect),
+		csrf.HashOIDCNonce(), extraParams)
 
 	csrfCookie, err := csrf.SetCookie(rw, req)
 	if err != nil {
@@ -803,7 +799,7 @@ func (p *OAuthProxy) doOAuthStart(rw http.ResponseWriter, req *http.Request, ove
 	}
 	// There are a lot of issues opened complaining about missing CSRF cookies.
 	// Try to log the INs and OUTs of OAuthProxy, to be easier to analyse these issues.
-	logger.Println("CSRF cookie %s was sent in request (state=%s).", csrfCookie.Name, hashOAuthState)
+	logger.Printf("CSRF cookie %v was sent in request (state=%v).", csrfCookie.Name, hashOAuthState)
 
 	http.Redirect(rw, req, loginURL, http.StatusFound)
 }
@@ -843,7 +839,7 @@ func (p *OAuthProxy) OAuthCallback(rw http.ResponseWriter, req *http.Request) {
 	if err != nil {
 		// There are a lot of issues opened complaining about missing CSRF cookies.
 		// Try to log the INs and OUTs of OAuthProxy, to be easier to analyse these issues.
-		cookies.LoggingCSRFCookiesInOAuthCallback(req, cookieName)
+		LoggingCSRFCookiesInOAuthCallback(req, cookieName)
 		logger.Println(req, logger.AuthFailure, "Invalid authentication via OAuth2: unable to obtain CSRF cookie: %s (state=%s)", err, nonce)
 		p.ErrorPage(rw, req, http.StatusForbidden, err.Error(), "Login Failed: Unable to find a valid CSRF token. Please try again.")
 		return
@@ -1251,4 +1247,28 @@ func (p *OAuthProxy) errorJSON(rw http.ResponseWriter, code int) {
 	// we need to send some JSON response because we set the Content-Type to
 	// application/json
 	rw.Write([]byte("{}"))
+}
+
+// LoggingCSRFCookiesInOAuthCallback Log all CSRF cookies found in HTTP request OAuth callback,
+// which were successfully parsed
+func LoggingCSRFCookiesInOAuthCallback(req *http.Request, cookieName string) {
+	cookies := req.Cookies()
+	if len(cookies) == 0 {
+		logger.Println(req, logger.AuthFailure, "No cookies were found in OAuth callback.")
+	} else {
+		var anyFound = false
+		for i := range cookies {
+			var c = cookies[i]
+			if cookieName == c.Name {
+				logger.Println(req, logger.AuthFailure, "The CSRF cookie %s was found in OAuth callback.", c.Name)
+				anyFound = true
+			} else if strings.Contains(c.Name, "_csrf") {
+				logger.Println(req, logger.AuthFailure, "The CSRF cookie %s was found in OAuth callback, but it is not the expected one (%s).", c.Name, cookieName)
+				anyFound = true
+			}
+		}
+		if !anyFound {
+			logger.Println(req, logger.AuthFailure, "Cookies were found in OAuth callback, but none was a CSRF cookie.")
+		}
+	}
 }
